refactor(ui-dropdown-item): extract styles and tokens into ui-dropdown-item.styles.ts
Split 712-line file into ui-dropdown-item.ts (437 lines) and ui-dropdown-item.styles.ts (283 lines). No behavioral changes.
